+++ src/main/java/com/sixt/service/framework/rpc/RpcClient.java
@@ -21,6 +21,7 @@ import com.sixt.service.framework.json.JsonRpcResponse;
 import com.sixt.service.framework.protobuf.ProtobufRpcRequest;
 import com.sixt.service.framework.protobuf.ProtobufRpcResponse;
 import com.sixt.service.framework.protobuf.ProtobufUtil;
+import com.sixt.service.framework.rpc.backoff.RetryBackOffFunction;
 
 import com.google.gson.JsonArray;
 import com.google.inject.Inject;
