+++ src/main/java/com/sixt/service/framework/rpc/backoff/TestRetryBackOff.java
@@ -1,0 +1,21 @@
+package com.sixt.service.framework.rpc.backoff;
+
+import java.time.Duration;
+
+public class TestRetryBackOff
+    implements RetryBackOffFunction {
+
+    @Override
+    public Duration timeout(int retryCounter) {
+        return null;
+    }
+
+    @Override
+    public void execute(int retryCounter) {
+        try {
+            Thread.currentThread().sleep(timeout(retryCounter).toMillis());
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+    }
+}
